Added 8 and 16bit spim stacks

+++ StackRegistration.h
@@ -10,15 +10,14 @@ class SpimStack;
 
 struct Threshold
 {
-	unsigned short	min;
-	unsigned short	max;
+	double			min;
+	double			max;
 
 	double			mean;
 	double			stdDeviation;
 
-	inline unsigned short getSpread() const { return max - min; }
-
-	inline float getRelativeValue(unsigned short v) const { return float(v - min) / float(max - min); }
+	inline double getSpread() const { return max - min; }
+	inline double getRelativeValue(double v) const { return (v - min) / (max - min); }
 };
 
 class ReferencePoints : boost::noncopyable
